authentification: add base for account management
New huge step for RORI. Migrate the whole communication layer from
scratch to something based on the RING project. Now, the whole auth
and communication between endpoints and RORI will use RING.

Allow me to concentrate on the core and use the work I do the rest
of the week. RING allow me to share files, do SIP calls, text
messages and manages accounts via the RING daemon.

This will need to defines a lot of things, because users and multi
devices has not the same signification between RING and RORI. I'll
describe this in the Wiki.